qsm/backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.25.0
httpx>=0.27.0
jinja2>=3.1.0
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
aiohttp>=3.9.0
redis>=5.0.0
